Rename pipeline_sensor -> run_status_sensor_definition
