+++ examples/hacker_news_assets/hacker_news_assets/sensors/slack_on_pipeline_failure_sensor.py
@@ -2,7 +2,7 @@ import os
 from typing import Dict, List
 
 from dagster import SensorDefinition
-from dagster.core.definitions.pipeline_definition_sensor import PipelineFailureSensorContext
+from dagster.core.definitions.run_status_sensor_definition import PipelineFailureSensorContext
 from dagster_slack import make_slack_on_pipeline_failure_sensor
 from hacker_news_assets.utils.slack_message import build_slack_message_blocks
