+++ python_modules/dagster-test/dagster_test/graph_job_op_toys/sensors.py
@@ -2,7 +2,7 @@ import os
 
 from dagster import AssetKey, RunRequest, SkipReason, check, sensor
 from dagster.core.definitions.decorators.sensor import asset_sensor
-from dagster.core.definitions.pipeline_definition_sensor import (
+from dagster.core.definitions.run_status_sensor_definition import (
     RunFailureSensorContext,
     run_failure_sensor,
 )
