+++ python_modules/dagster/dagster_tests/core_tests/definitions_tests/test_pipeline_sensor.py
@@ -1,5 +1,5 @@
 from dagster import repository
-from dagster.core.definitions.pipeline_definition_sensor import pipeline_failure_sensor
+from dagster.core.definitions.run_status_sensor_definition import pipeline_failure_sensor
 
 
 def test_pipeline_failure_sensor_def():
