+++ python_modules/dagster/dagster_tests/daemon_tests/test_types.py
@@ -1,4 +1,4 @@
-from dagster.core.definitions.pipeline_definition_sensor import RunStatusSensorCursor
+from dagster.core.definitions.run_status_sensor_definition import RunStatusSensorCursor
 from dagster.daemon.types import DaemonHeartbeat
 from dagster.serdes import deserialize_as, deserialize_json_to_dagster_namedtuple
 from dagster.utils.error import SerializableErrorInfo
