+++ python_modules/libraries/dagster-msteams/dagster_msteams/sensors.py
@@ -1,6 +1,6 @@
 from typing import Callable, Optional
 
-from dagster.core.definitions.pipeline_definition_sensor import (
+from dagster.core.definitions.run_status_sensor_definition import (
     PipelineFailureSensorContext,
     pipeline_failure_sensor,
 )
